# webdoc/docs/project_specs/007-api-layer-crud/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Focal CRUD API
  description: Configuration management endpoints for Focal cognitive engine
  version: 1.0.0
servers:
  - url: /v1
    description: API v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with tenant_id, sub (user_id), and roles claims
  schemas:
...
    RuleCreate:
      type: object
      required: [name, condition_text, action_text]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        condition_text:
          type: string
          minLength: 1
        action_text:
          type: string
          minLength: 1
        scope:
          type: string
          enum: [global, scenario, step]
          default: global
        scope_id:
          type: string
          format: uuid
        priority:
          type: integer
          minimum: -100
          maximum: 100
          default: 0
        enabled:
          type: boolean
          default: true
        max_fires_per_session:
          type: integer
          default: 0
        cooldown_turns:
          type: integer
          default: 0
        attached_tool_ids:
          type: array
          items:
            type: string

    # Bulk Operations
    BulkOperation:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [create, update, delete]
        id:
          type: string
          format: uuid
        data:
          type: object
    BulkRequest:
      type: object
      required: [operations]
      properties:
        operations:
          type: array
          items:
            $ref: '#/components/schemas/BulkOperation'
          maxItems: 50
    BulkResult:
      type: object
      required: [action, success]
      properties:
        action:
          type: string
        success:
          type: boolean
        id:
          type: string
          format: uuid
        error:
          type: string
    BulkResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkResult'

    # Templates
...
paths:
...
  /agents/{agent_id}/rules:
    parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List rules
      tags: [Rules]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: scope
          in: query
          schema:
            type: string
            enum: [global, scenario, step]
        - name: scope_id
          in: query
          schema:
            type: string
            format: uuid
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: priority_gte
          in: query
          schema:
            type: integer
        - name: priority_lte
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated list of rules
    post:
      summary: Create rule
      tags: [Rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
      responses:
        '201':
          description: Rule created
  /agents/{agent_id}/rules/bulk:
    parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Bulk rule operations
      tags: [Rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRequest'
      responses:
        '200':
          description: Bulk operation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
  /agents/{agent_id}/rules/{rule_id}:
    parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: rule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get rule
      tags: [Rules]
      responses:
        '200':
          description: Rule details
    put:
      summary: Update rule
      tags: [Rules]
      responses:
        '200':
          description: Rule updated
    delete:
      summary: Delete rule
      tags: [Rules]
      responses:
        '204':
          description: Rule deleted

  # Templates
...